+++ src/mygame/Tower.java
@@ -4,7 +4,6 @@
  */
 package mygame;
 
-import com.jme3.asset.AssetManager;
 import com.jme3.math.Quaternion;
 import com.jme3.math.Vector3f;
 import com.jme3.renderer.RenderManager;
@@ -17,13 +16,13 @@ import com.jme3.scene.control.AbstractControl;
  *
  * @author Amir, Daniel, Lukas
  */
-class Tower extends ABstractTowerControl{
+class Tower extends AbstractControl {
 
     public static final int TYPE_MARINE = 0;
     public static final int TYPE_POLICE = 1;
     public static final int TYPE_UNICORN = 2;
     public static final int TYPE_MURICA = 3;
-   
+    private Spatial geom;
     private int cooldown = 1000;
     private long startTime;
     private float visibility = 40f;
@@ -33,9 +32,8 @@ class Tower extends ABstractTowerControl{
     public int type = 0;
     private Quaternion q;
     private IllegalImmigrant target;
-    private Spatial geom;
-      private AssetManager assetManager;
-  /*  public Tower(int type) {
+
+    public Tower(int type) {
         this.type = type;
 
         if (type == TYPE_UNICORN) {
@@ -57,44 +55,75 @@ class Tower extends ABstractTowerControl{
     /**
      * für die Vorschau bevor der Turm platziert wird
      *
-     *//*
+     */
     public Geometry createGeometry() {
         return GeometryCreator.instance.createBox(Vector3f.ZERO);
     }
-*/
-  
 
     @Override
-    public void init(Vector3f position) {
-        this.assetManager = Main.instance.getAssetManager();
-       geom = createGeometry();
-                geom.rotate((float) Math.toRadians(90), 0, 0);
-        geom.setLocalTranslation(position);
-        Main.instance.attachToRootNode(geom);
-        startTime = System.currentTimeMillis();
-       // s = Main.instance.getSpawner();
-        geom.addControl(this);
+    protected void controlUpdate(float tpf) {
+        if (target == null) {
+            target = getNearestImmigrant();
+        }
+        if (target != null) {
+            spatial.lookAt(target.getPosition(), new Vector3f(0, 0, 1));
+
+
+            if (target.getPosition().distance(spatial.getLocalTranslation()) <= range) {
+                if (System.currentTimeMillis() - startTime >= cooldown) {
+
+                    if (target != null) {
+                        new Projectile(target, this, projectileType);
+                        target = getNearestImmigrant();
+
+                        startTime = System.currentTimeMillis();
+                    }
+                }
+            }
+        }
     }
-  public Spatial createGeometry() {
-     Spatial geome =  assetManager.loadModel("Models/unicornfuv2.j3o");
-        return   geome;
-    }
-  
+
     @Override
-     void getProjectile() {
-       new Projectile(target, this, getProjectileType());
-        
+    protected void controlRender(RenderManager rm, ViewPort vp) {
     }
 
-    @Override
-     int getProjectileType() {
-       return 0;
+    public IllegalImmigrant getNearestImmigrant() {
+        float distance = -1.0f;
+        IllegalImmigrant nearest = null;
+        for (Object o : s.getImmigrants()) {
+            IllegalImmigrant i = (IllegalImmigrant) o;
+            float d = getPosition().distance(i.getPosition());
+            switch (type) {
+                case TYPE_MARINE:
+                    if (d < visibility && (d < distance || distance == -1) && (i.targeted == false)) {
+                        nearest = i;
+                        distance = d;
+                    }
+                    break;
+                case TYPE_POLICE:
+                    if (d < visibility && (d < distance || distance == -1)) {
+                        if (i.getTaserTicks() == 0) {
+                            nearest = i;
+                            distance = d;
+                        }
+                    }
+                    break;
+                case TYPE_UNICORN:
+                    if (d < visibility && (d < distance || distance == -1) && (i.targeted == false)) {
+                        nearest = i;
+                        distance = d;
+                    }
+                    break;
+            }
+        }
+
+        if (nearest != null) {
+            nearest.targeted = true;
+        }
+        return nearest;
     }
 
-
-
-
-
-
-   
+    public Vector3f getPosition() {
+        return geom.getLocalTranslation();
+    }
 }
